backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
kafka-python
pymysql
cryptography
mysql-connector-python
confluent-kafka